backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "imagehash>=4.3.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "scikit-image>=0.25.2",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app/"]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
]
